langchain-poc/openlit/requirements.txt
langchain
langchain-community
langchain-ollama
opentelemetry-sdk
opentelemetry-api
opentelemetry-semantic-conventions
python-dotenv
openlit
botocore
setuptools
boto3
aws-opentelemetry-distro-genai-beta==0.1.0b6
fastapi
uvicorn[standard]
